back/requirements.txt
# Complete requirements matching your versions
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
channels==4.2.2
channels-redis==4.2.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
gunicorn==23.0.0
h11==0.16.0
kombu==5.5.3
msgpack==1.1.0
packaging==25.0
Pillow==11.2.1
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
qrcode[pil]>=7.4
redis==6.1.0
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0





